done with the refactor
when everyone's tests get merged I'll fix it up so it passes CI
