+++ public/src/actions/admin_contact_lists.jsx
@@ -1,0 +1,25 @@
+import axios from 'axios';
+import { SET_CAMPAIGN_FORM_CONTACT_LIST } from '../reducers/admin_contact_lists';
+
+export function setContactListOptions(contactLists) {
+  return {
+    type: SET_CAMPAIGN_FORM_CONTACT_LIST,
+    payload: contactLists
+  };
+}
+
+export function fetchAllContactLists() {
+  return dispatch => axios.get('/contactLists')
+    .then((res) => {
+      const contactLists = res.data;
+      dispatch(setContactListOptions(contactLists));
+      return contactLists;
+    })
+    .catch((err) => {
+      const customError = {
+        message: `error fetching contact lists: ${err}`,
+        name: 'contact lists get request from campaignPage component'
+      };
+      throw customError;
+    });
+}
