+++ public/src/actions/admin_questions.jsx
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import { destroy } from 'redux-form';
 import { SET_QUESTIONS, SET_QUESTION_CURRENT } from '../reducers/admin_questions';
 
 export function setQuestionList(questionList) {
@@ -24,6 +25,37 @@ export function fetchAllQuestions() {
     return dispatch(setQuestionList(questionsList));
   })
   .catch((err) => {
-    console.log('error fetching all questions: ', err);
+    const customError = {
+      message: `error fetching questions: ${err}`,
+      name: 'question get request from script_form component'
+    };
+    throw customError;
+  });
+}
+
+export default function createQuestion(questionInfo, history) {
+  const { title, description, type } = questionInfo;
+  // Create single string for all options
+  const keys = Object.keys(questionInfo);
+  const options = keys.filter(key => key.indexOf('option') >= 0);
+  const responses = options.join(',');
+
+  return dispatch => axios.post('/questions', {
+    title,
+    description,
+    type,
+    responses
+  })
+  .then((res) => {
+    history.goBack();
+    dispatch(destroy('CreateQuestion'));
+    return res;
+  })
+  .catch((err) => {
+    const customError = {
+      message: `error in creating new questions to db: ${err}`,
+      name: 'createQuestion function error in actions/questions.jsx'
+    };
+    throw customError;
   });
 }
