+++ public/src/containers/create_question.jsx
@@ -2,7 +2,7 @@ import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { reduxForm, formValueSelector, destroy } from 'redux-form';
 import CreateQuestion from '../components/question/create_question';
-import createQuestion from '../actions/questions';
+import createQuestion from '../actions/admin_questions';
 import validate from '../helpers/question_validation';
 
 const FORM_NAME = 'CreateQuestion';
