--- public/src/reducers/script_form.jsx
@@ -1,16 +1,0 @@
-const questions = [];
-
-export const SET_SCRIPT_FORM_QUESTION_OPTIONS = 'SET_SCRIPT_FORM_QUESTION_OPTIONS';
-
-export function questionOptionsReducer(state = questions, action) {
-  const { type, payload } = action;
-  switch (type) {
-    case SET_SCRIPT_FORM_QUESTION_OPTIONS:
-      return {
-        ...state,
-        questionOptions: payload
-      };
-    default:
-      return state;
-  }
-}
