+++ public/src/store.jsx
@@ -10,20 +10,18 @@ import { createLogger } from 'redux-logger';
 import { authStatusReducer, LOGOUT_USER } from './reducers/login';
 import { accountInfoReducer } from './reducers/account_info';
 import { campaignListReducer } from './reducers/campaign';
-import { campaignFormReducer } from './reducers/create_campaign';
-import { questionOptionsReducer } from './reducers/script_form';
 import { adminQuestionsReducer } from './reducers/admin_questions';
 import { adminScriptsReducer } from './reducers/admin_scripts';
+import { adminContactListsReducer } from './reducers/admin_contact_lists';
 
 const appReducer = combineReducers({
   form: formReducer,
   auth: authStatusReducer,
   account_info: accountInfoReducer,
   admin_campaigns: campaignListReducer,
-  campaign_form: campaignFormReducer,
   admin_questions: adminQuestionsReducer,
   admin_scripts: adminScriptsReducer,
-  script_form: questionOptionsReducer
+  admin_contact_lists: adminContactListsReducer
 });
 
 const rootReducer = (state, action) => {
@@ -37,5 +35,8 @@ const middleware = [immutable(), createLogger(), promise(), thunk];
 const store = createStore(rootReducer, compose(applyMiddleware(...middleware), autoRehydrate()));
 
 persistStore(store);
+
+// to reset the store:
+// persistStore(store).purge();
 
 export default store;
